pyproject.toml
[tool.poetry]
name = "crawler-api"
version = "0.1.0"
description = ""
authors = ["Henrique Boneto"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
httpx = "^0.27.0"
google-cloud-pubsub = "^2.19.4"
google-cloud-firestore = "^2.15.0"
aiohttp = "^3.9.3"
pyppeteer = "^2.0.0"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.5"
pytest = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
